+++ src/checkout/main.go
@@ -364,19 +364,19 @@ func (cs *checkout) PlaceOrder(ctx context.Context, req *pb.PlaceOrderRequest) (
 	totalPriceFloat, _ := strconv.ParseFloat(fmt.Sprintf("%d.%02d", total.GetUnits(), total.GetNanos()/1000000000), 64)
 
 	span.SetAttributes(
-		attribute.String("app.order.id", orderID.String()),
+		attribute.String("demo.order.id", orderID.String()),
 		attribute.Float64("app.shipping.amount", shippingCostFloat),
-		attribute.Float64("app.order.amount", totalPriceFloat),
-		attribute.Int("app.order.items.count", len(prep.orderItems)),
+		attribute.Float64("demo.order.amount", totalPriceFloat),
+		attribute.Int("demo.order.items.count", len(prep.orderItems)),
 		shippingTrackingAttribute,
 	)
 	logger.LogAttrs(
 		ctx,
 		slog.LevelInfo, "order placed",
-		slog.String("app.order.id", orderID.String()),
+		slog.String("demo.order.id", orderID.String()),
 		slog.Float64("app.shipping.amount", shippingCostFloat),
-		slog.Float64("app.order.amount", totalPriceFloat),
-		slog.Int("app.order.items.count", len(prep.orderItems)),
+		slog.Float64("demo.order.amount", totalPriceFloat),
+		slog.Int("demo.order.items.count", len(prep.orderItems)),
 		slog.String("app.shipping.tracking.id", shippingTrackingID),
 	)
 
@@ -437,8 +437,8 @@ func (cs *checkout) prepareOrderItemsAndShippingQuoteFromCart(ctx context.Contex
 
 	span.SetAttributes(
 		attribute.Float64("app.shipping.amount", shippingCostFloat),
-		attribute.Int("app.cart.items.count", int(totalCart)),
-		attribute.Int("app.order.items.count", len(orderItems)),
+		attribute.Int("demo.cart.items.count", int(totalCart)),
+		attribute.Int("demo.order.items.count", len(orderItems)),
 	)
 	return out, nil
 }